nfp/pyproject.toml
[tool.poetry]
name = "nfp-backend"
version = "0.1.0"
description = ""
authors = ["Roger Hendricks <hendricks.r@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
alembic = "^1.8.1"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
cffi = "^1.15.1"
cryptography = "^38.0.4"
databases = "^0.6.2"
SQLAlchemy = "1.4.41"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
